third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/source-list-parsing-02.html

<!DOCTYPE html>
<html>
<head>
<script src='resources/multiple-iframe-test.js'></script>
<script>
var tests = [
    ['yes', 'script-src\thttp://127.0.0.1:8000', 'resources/script.js'],
    ['yes', 'script-src http://127.0.0.1:8000   \t ', 'resources/script.js'],
    ['yes', 'script-src http://127.0.0.1:* ', 'resources/script.js'],
    ['yes', 'script-src \'self\' ;', 'resources/script.js'],
];
</script>
</head>
<body onload="test()">
  <p>
    These scripts should execute even though there are parse errors in the policy.
  </p>
