third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/source-list-parsing-02-expected.txt

These scripts should execute even though there are parse errors in the policy.

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
PASS

--------
Frame: '<!--framePath //<!--frame2-->-->'
--------
PASS

--------
Frame: '<!--framePath //<!--frame3-->-->'
--------
PASS
